% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaboost.R
\name{adaboost}
\alias{adaboost}
\title{Perform Classic AdaBoost}
\usage{
adaboost(x, yname, treedepth = 3, event = 1, m = 50)
}
\arguments{
\item{x}{A dataframe.}

\item{yname}{A character vector indicating the column name of the binary target.}

\item{treedepth}{The maximal depth of the tree.}

\item{event}{Value that indicates the even in `y`.}

\item{m}{The number of trees.}
}
\value{
A list containing:
* The predicted values of x.
* The confidence of the predictions.
* A list with the `m` trees.
* The vector with alpha values.
* The column name of `y`.
}
\description{
This is the implementation of the classic AdaBoost algorithm.
}
\details{
This is the AdaBoost implementation as described on page 258 of
the articel.
}
\examples{
adaboost(mtcars, "am")
}
\references{
Joshi, Kumar, Agarwal 2001
}

